% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_accident}
\alias{accident_accident}
\title{Data by State for some accidents
The data set included to test importinting the data in to a package}
\format{a data frame with columns
\describe{
\item{STATE}{state}
\item{COUNTY}{county}
\item{DAY}{day}
\item{many_more_other_items}{some other info}
}}
\source{
Coursera
}
\usage{
accident_accident
}
\description{
Data by State for some accidents
The data set included to test importinting the data in to a package
}
\examples{
\dontrun{
accident_accident
}
}
\keyword{datasets}
