% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents by sate by year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a numeric value normally between 1 and 52. The value is validated against unique values in the dataframe}

\item{year}{number or string that can be coerced to a number}
}
\value{
This function produces a plot and returns NULL
}
\description{
This function takes state number and a year and creates a plot of accidents overlayed over a polygon of a state
}
\examples{
\dontrun{
fars_map_state(10,2014)
}

}
