% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Simple function that takes a file name as an input and reads a .csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string of a file name}
}
\value{
The function returns a dataframe. No modifications to the source file
}
\description{
The file has to be present in the source directory. If a file is not present, the function stops execution
}
