% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read files from the source directory based on a vector of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of years to be included in to a file name}
}
\value{
This function returns a list with each element containing data for each year in the input
}
\description{
The function takes a vector of years and reads files with each year in the name into a list
}
\examples{
\dontrun{
df<-fars_read_years(c(2014,2015,2016))
}
}
