% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Produce summary counts from the specified years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of years to be included in to a file name}
}
\value{
a dataframe that counts number of records by year, by month
}
\description{
This function takes a vector of years, load the files based on the template from make_file name and counts records
by year, by month
}
\examples{
\dontrun{
df2<-fars_summarize_years(c(2014,2015,2016))
}

}
