% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Simple function that takes a year as a number and merges it within the template accident_YEAR.csv.bz2}
\usage{
make_filename(year)
}
\arguments{
\item{year}{number or string that can be coerced to a number}
}
\value{
The function returns a string that will be used as file input
}
\description{
Simple function that takes a year as a number and merges it within the template accident_YEAR.csv.bz2
}
